--- src/main/java/com/Project/reserve/ReserveDao.java
@@ -1,7 +1,0 @@
-package com.Project.reserve;
-
-public interface ReserveDao {
-
-	int insertReserve(ReserveModel reserveModel);
-
-}
