--- src/main/java/com/Project/reserve/ReserveService.java
@@ -1,19 +1,0 @@
-package com.Project.reserve;
-
-import javax.annotation.Resource;
-
-import org.mybatis.spring.SqlSessionTemplate;
-import org.springframework.stereotype.Service;
-
-@Service
-public class ReserveService implements ReserveDao {
-
-	@Resource
-	private SqlSessionTemplate sqlSessionTemplate;
-
-	@Override
-	public int insertReserve(ReserveModel reserveModel) {
-		return sqlSessionTemplate.insert("reserve.insertReserve", reserveModel);
-	}
-
-}
